+++ src/libs/safe/safe.test.ts
@@ -1,9 +1,89 @@
-import { describe, expect, test } from '@jest/globals'
+import { describe, expect, jest, test } from '@jest/globals'
+import { getAddress } from 'ethers'
 
+import { Hex } from '../../interfaces/hex'
 import { buildSafeMessageOrigin, parseSafeMessageOrigin } from './helpers'
-import { normalizeSafeGlobalMessage } from './safe'
+import { getSafeAccountByOwner, normalizeSafeGlobalMessage } from './safe'
 
 import type { EIP712TypedData } from '@safe-global/types-kit'
+
+const OWNER = '0xD8293ad21678c6F09Da139b4B62D38e514a03B78' as Hex
+const OTHER_OWNER = '0x94b0080A00579C1307B0eF2C499AD98A8ce58e58'
+const SAFE_ADDRESS = '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48'
+
+const getSafeInfo = (owners = [OWNER]) => ({
+  address: SAFE_ADDRESS,
+  fallbackHandler: '0x0000000000000000000000000000000000000000',
+  guard: '0x0000000000000000000000000000000000000000',
+  masterCopy: '0x0000000000000000000000000000000000000000',
+  modules: [],
+  nonce: 0,
+  owners,
+  threshold: 1,
+  version: '1.4.1'
+})
+
+const getSafeCreationInfo = () => ({
+  created: '2025-01-01T00:00:00Z',
+  creator: OWNER,
+  factoryAddress: '0x1234567890123456789012345678901234567890',
+  saltNonce: '1',
+  setupData: '0x1234',
+  singleton: '0x2345678901234567890123456789012345678901',
+  transactionHash: `0x${'1'.repeat(64)}`
+})
+
+const createApi = (owners = [OWNER]) => ({
+  getSafeCreationInfo: jest.fn(async () => getSafeCreationInfo()),
+  getSafeInfo: jest.fn(async () => getSafeInfo(owners))
+})
+
+describe('getSafeAccountByOwner', () => {
+  test('falls back to the next deployed network when fetching Safe details fails', async () => {
+    const mainnetApi = createApi()
+    mainnetApi.getSafeInfo.mockRejectedValue(new Error('Service unavailable'))
+    const optimismApi = createApi()
+    const apiKitFactory = jest.fn((chainId: bigint) => (chainId === 1n ? mainnetApi : optimismApi))
+
+    const result = await getSafeAccountByOwner(SAFE_ADDRESS, OWNER, [1n, 10n], apiKitFactory)
+
+    expect(mainnetApi.getSafeInfo).toHaveBeenCalledWith(SAFE_ADDRESS)
+    expect(mainnetApi.getSafeCreationInfo).not.toHaveBeenCalled()
+    expect(optimismApi.getSafeInfo).toHaveBeenCalledWith(SAFE_ADDRESS)
+    expect(optimismApi.getSafeCreationInfo).toHaveBeenCalledWith(SAFE_ADDRESS)
+    expect(result.account).toMatchObject({
+      addr: getAddress(SAFE_ADDRESS),
+      associatedKeys: [OWNER],
+      deployedOn: [1n, 10n]
+    })
+    expect(result.failed).toBe(false)
+  })
+
+  test('falls back to the next deployed network when the first one does not include the owner', async () => {
+    const mainnetApi = createApi([OTHER_OWNER])
+    const optimismApi = createApi()
+    const apiKitFactory = jest.fn((chainId: bigint) => (chainId === 1n ? mainnetApi : optimismApi))
+
+    const result = await getSafeAccountByOwner(SAFE_ADDRESS, OWNER, [1n, 10n], apiKitFactory)
+
+    expect(mainnetApi.getSafeInfo).toHaveBeenCalledWith(SAFE_ADDRESS)
+    expect(mainnetApi.getSafeCreationInfo).not.toHaveBeenCalled()
+    expect(optimismApi.getSafeInfo).toHaveBeenCalledWith(SAFE_ADDRESS)
+    expect(optimismApi.getSafeCreationInfo).toHaveBeenCalledWith(SAFE_ADDRESS)
+    expect(result.account?.addr).toBe(getAddress(SAFE_ADDRESS))
+    expect(result.failed).toBe(false)
+  })
+
+  test('does not return an account when no deployed network includes the owner', async () => {
+    const api = createApi([OTHER_OWNER])
+
+    const result = await getSafeAccountByOwner(SAFE_ADDRESS, OWNER, [1n, 10n], () => api)
+
+    expect(api.getSafeInfo).toHaveBeenCalledTimes(2)
+    expect(api.getSafeCreationInfo).not.toHaveBeenCalled()
+    expect(result).toEqual({ account: null, failed: false })
+  })
+})
 
 describe('normalizeSafeGlobalMessage', () => {
   test('converts a typed message domain chainId bigint to a decimal string', () => {
