+++ src/libs/safe/safe.ts
@@ -17,10 +17,12 @@ import { SignTypedDataVersion, TypedDataUtils } from '@metamask/eth-sig-util'
 import SafeApiKit from '@safe-global/api-kit'
 
 import SafeAbi from '../../../contracts/compiled/Safe.json'
+import { SAFE_API_TIMEOUT_MS } from '../../consts/safe'
 import { Hex } from '../../interfaces/hex'
 import { RPCProvider } from '../../interfaces/provider'
-import { SafeTx } from '../../interfaces/safe'
+import { SafeAccountByOwner, SafeTx } from '../../interfaces/safe'
 import { CallsUserRequest, TypedMessageUserRequest } from '../../interfaces/userRequest'
+import { withTimeout } from '../../utils/with-timeout'
 import wait from '../../utils/wait'
 import { adaptTypedMessageForMetaMaskSigUtil } from '../signMessage/signMessage'
 import { decodeMultiSend, multiCallAbi, parseSafeMessageOrigin } from './helpers'
@@ -60,6 +62,77 @@ export function getApiKit(chainId: bigint) {
     apiKey: process.env.SAFE_API_KEY,
     txServiceUrl: getTxServiceUrl(chainId)
   })
+}
+
+type SafeAccountApiKitFactory = (
+  chainId: bigint
+) => Pick<ReturnType<typeof getApiKit>, 'getSafeInfo' | 'getSafeCreationInfo'>
+
+export async function getSafeAccountByOwner(
+  safeAddr: string,
+  owner: Hex,
+  deployedOn: bigint[],
+  apiKitFactory: SafeAccountApiKitFactory = getApiKit
+): Promise<{ account: SafeAccountByOwner | null; failed: boolean }> {
+  const getAccountFromChain = async (
+    [chainId, ...remainingChainIds]: bigint[],
+    hasRequestFailed = false
+  ): Promise<{
+    account: SafeAccountByOwner | null
+    failed: boolean
+  }> => {
+    if (chainId === undefined) return { account: null, failed: hasRequestFailed }
+
+    const apiKit = apiKitFactory(chainId)
+    try {
+      const safeInfo = await withTimeout(() => apiKit.getSafeInfo(safeAddr), {
+        timeoutMs: SAFE_API_TIMEOUT_MS,
+        message: `Safe API: get Safe info timed out after ${SAFE_API_TIMEOUT_MS}ms`
+      })
+      const address = getAddress(safeAddr.toLowerCase())
+      const owners = safeInfo.owners.map((safeOwner: string) => getAddress(safeOwner.toLowerCase()))
+      if (!owners.some((safeOwner) => safeOwner === owner)) {
+        return getAccountFromChain(remainingChainIds, hasRequestFailed)
+      }
+
+      const safeCreationInfo = await withTimeout(() => apiKit.getSafeCreationInfo(safeAddr), {
+        timeoutMs: SAFE_API_TIMEOUT_MS,
+        message: `Safe API: get Safe creation info timed out after ${SAFE_API_TIMEOUT_MS}ms`
+      })
+
+      return {
+        account: {
+          addr: address,
+          associatedKeys: owners,
+          initialPrivileges: owners.map((safeOwner) => [safeOwner, '0x01']),
+          creation: null,
+          safeCreation: {
+            factoryAddr: safeCreationInfo.factoryAddress as Hex,
+            singleton: safeCreationInfo.singleton as Hex,
+            setupData: safeCreationInfo.setupData as Hex,
+            saltNonce: safeCreationInfo.saltNonce
+              ? (toBeHex(BigInt(safeCreationInfo.saltNonce), 32) as Hex)
+              : (toBeHex(0, 32) as Hex),
+            version: safeInfo.version
+          },
+          preferences: {
+            label: 'Safe',
+            pfp: address
+          },
+          deployedOn
+        },
+        failed: false
+      }
+    } catch (error) {
+      console.error(
+        `Failed to retrieve Safe account ${safeAddr} on network ${chainId.toString()}`,
+        error
+      )
+      return getAccountFromChain(remainingChainIds, true)
+    }
+  }
+
+  return getAccountFromChain(deployedOn)
 }
 
 export async function getCalculatedSafeAddress(
